+++ packages/datamanager-firestore-mobx/src/index.ts
@@ -1,5 +1,5 @@
-export * from '@data-weave/backend-firestore/src'
-export * from '@data-weave/datamanager/src'
+export * from '@data-weave/backend-firestore'
+export * from '@data-weave/datamanager'
 
 export * from './ObservableFirestoreList'
 export * from './ObservableFirestoreReference'