+++ packages/backend-firestore/CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.27](https://github.com/data-weave/datamanager/compare/v0.4.26...v0.4.27) (2025-11-10)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.26](https://github.com/data-weave/datamanager/compare/v0.4.25...v0.4.26) (2025-11-10)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.25](https://github.com/data-weave/datamanager/compare/v0.4.24...v0.4.25) (2025-11-10)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.24](https://github.com/data-weave/datamanager/compare/v0.4.23...v0.4.24) (2025-11-10)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.23](https://github.com/data-weave/datamanager/compare/v0.4.22...v0.4.23) (2025-11-10)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.22](https://github.com/data-weave/datamanager/compare/v0.4.21...v0.4.22) (2025-11-09)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.21](https://github.com/data-weave/datamanager/compare/v0.4.20...v0.4.21) (2025-11-09)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.20](https://github.com/data-weave/datamanager/compare/v0.4.19...v0.4.20) (2025-11-09)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.19](https://github.com/data-weave/datamanager/compare/v0.4.18...v0.4.19) (2025-11-08)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.18](https://github.com/data-weave/datamanager/compare/v0.4.17...v0.4.18) (2025-11-08)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.17](https://github.com/data-weave/datamanager/compare/v0.4.16...v0.4.17) (2025-11-08)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.16](https://github.com/data-weave/datamanager/compare/v0.4.15...v0.4.16) (2025-11-08)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.15](https://github.com/data-weave/datamanager/compare/v0.4.13...v0.4.15) (2025-11-08)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.14](https://github.com/data-weave/datamanager/compare/v0.4.13...v0.4.14) (2025-11-08)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.13](https://github.com/data-weave/datamanager/compare/v0.4.8...v0.4.13) (2025-11-08)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.8](https://github.com/data-weave/datamanager/compare/v0.4.4...v0.4.8) (2025-10-22)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.7](https://github.com/data-weave/datamanager/compare/v0.4.4...v0.4.7) (2025-10-22)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.6](https://github.com/data-weave/datamanager/compare/v0.4.4...v0.4.6) (2025-10-22)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
+## [0.4.5](https://github.com/data-weave/datamanager/compare/v0.4.4...v0.4.5) (2025-10-22)
+
+**Note:** Version bump only for package @data-weave/backend-firestore
+
 ## [0.4.4](https://github.com/data-weave/datamanager/compare/v0.4.3...v0.4.4) (2025-10-15)
 
 **Note:** Version bump only for package @data-weave/backend-firestore